# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ services/web/app/coffee/Features/Uploads/ProjectUploadController.coffee
@@ -30,6 +30,9 @@ module.exports = ProjectUploadController =
 		{name, path} = req.files.qqfile
 		project_id   = req.params.Project_id
 		folder_id    = req.query.folder_id
+		if !name? or name.length == 0 or name.length > 150
+			logger.err project_id:project_id, name:name, "bad name when trying to upload file"
+			return res.send success: false
 		FileSystemImportManager.addEntity project_id, folder_id, name, path, true, (error, entity) ->
 			fs.unlink path, ->
 			timer.done()
